🛜 Simplify wifi services
