--- esp32/lib/ESP32-sveltekit/StatefulService.cpp
@@ -1,18 +1,0 @@
-/**
- *   ESP32 SvelteKit
- *
- *   A simple, secure and extensible framework for IoT projects for ESP32 platforms
- *   with responsive Sveltekit front-end built with TailwindCSS and DaisyUI.
- *   https://github.com/theelims/ESP32-sveltekit
- *
- *   Copyright (C) 2018 - 2023 rjwats
- *   Copyright (C) 2023 theelims
- *
- *   All Rights Reserved. This software may be modified and distributed under
- *   the terms of the LGPL v3 license. See the LICENSE file for details.
- **/
-
-#include <StatefulService.h>
-
-update_handler_id_t StateUpdateHandlerInfo::currentUpdatedHandlerId = 0;
-hook_handler_id_t StateHookHandlerInfo::currentHookHandlerId = 0;
