--- esp32/lib/ESP32-sveltekit/WiFiScanner.h
@@ -1,41 +1,0 @@
-#ifndef WiFiScanner_h
-#define WiFiScanner_h
-
-/**
- *   ESP32 SvelteKit
- *
- *   A simple, secure and extensible framework for IoT projects for ESP32 platforms
- *   with responsive Sveltekit front-end built with TailwindCSS and DaisyUI.
- *   https://github.com/theelims/ESP32-sveltekit
- *
- *   Copyright (C) 2018 - 2023 rjwats
- *   Copyright (C) 2023 theelims
- *
- *   All Rights Reserved. This software may be modified and distributed under
- *   the terms of the LGPL v3 license. See the LICENSE file for details.
- **/
-
-#include <WiFi.h>
-
-#include <ArduinoJson.h>
-#include <PsychicHttp.h>
-#include <SecurityManager.h>
-
-#define SCAN_NETWORKS_SERVICE_PATH "/api/scanNetworks"
-#define LIST_NETWORKS_SERVICE_PATH "/api/listNetworks"
-
-class WiFiScanner {
-  public:
-    WiFiScanner(PsychicHttpServer *server, SecurityManager *securityManager);
-
-    void begin();
-
-  private:
-    PsychicHttpServer *_server;
-    SecurityManager *_securityManager;
-
-    esp_err_t scanNetworks(PsychicRequest *request);
-    esp_err_t listNetworks(PsychicRequest *request);
-};
-
-#endif // end WiFiScanner_h
